app/requirements.txt
binance-connector
sqlalchemy